removing redis1 functionality
