+++ links/redis7.py
@@ -1869,3 +1869,128 @@ def account_created(ip,username):
 	my_server = redis.Redis(connection_pool=POOL)
 	my_server.setex("ip:"+str(ip),username,FOUR_MINS)
 
+
+###################### legacy redis1 functions ##############################
+###################### ecomm.py  ############################################
+
+def add_exchange_visitor(user_id):
+	redis.Redis(connection_pool=POOL).sadd("ftux:"+str(user_id), '14')
+
+
+def first_time_exchange_visitor(user_id):
+	if redis.Redis(connection_pool=POOL).sismember("ftux:"+str(user_id),'14'):
+		return False
+	else:
+		return True	
+
+
+def add_photo_ad_visitor(user_id):
+	my_server = redis.Redis(connection_pool=POOL)
+	set_name = "ftux:"+str(user_id)
+	my_server.sadd(set_name, '17')
+
+
+def first_time_photo_ads_visitor(user_id):
+	my_server = redis.Redis(connection_pool=POOL)
+	set_name = "ftux:"+str(user_id)
+	if my_server.sismember(set_name,'17'):
+		return False
+	else:
+		return True		
+
+################################## maint_views.py ##############################################
+
+
+def get_inactives(get_50K=False, get_10K=False, get_5K=False, key=None):
+	my_server = redis.Redis(connection_pool=POOL)
+	if not key:
+		key = "inactive_users"
+	if get_50K:
+		remaining = get_inactive_count(server=my_server,key_name = None if key == 'inactive_users' else key)
+		if remaining < 50000:
+			data = my_server.zrange(key,0,-1,withscores=True)
+			my_server.delete(key)
+			return data, True
+		else:
+			data = my_server.zrange(key,0,49999,withscores=True)
+			my_server.zremrangebyrank(key,0,49999)
+			return data, False
+	elif get_10K:
+		remaining = get_inactive_count(server=my_server,key_name = None if key == 'inactive_users' else key)
+		if remaining < 10000:
+			data = my_server.zrange(key,0,-1,withscores=True)
+			my_server.delete(key)
+			return data, True
+		else:
+			data = my_server.zrange(key,0,9999,withscores=True)
+			my_server.zremrangebyrank(key,0,9999)
+			return data, False
+	elif get_5K:
+		remaining = get_inactive_count(server=my_server,key_name = None if key == 'inactive_users' else key)
+		if remaining < 5000:
+			data = my_server.zrange(key,0,-1,withscores=True)
+			my_server.delete(key)
+			return data, True
+		else:
+			data = my_server.zrange(key,0,4999,withscores=True)
+			my_server.zremrangebyrank(key,0,4999)
+			return data, False
+	else:
+		return my_server.zrange(key,0,-1,withscores=True)
+
+def set_inactives(inactive_list):
+	if inactive_list:
+		redis.Redis(connection_pool=POOL).zadd("inactive_users", *inactive_list)
+
+
+def get_inactive_count(server=None, key_name=None):
+	if not server:
+		server = redis.Redis(connection_pool=POOL)
+	if not key_name:
+		return server.zcard("inactive_users")
+	else:
+		return server.zcard(key_name)
+
+def create_inactives_copy():
+	"""Creates a copy of the sorted set using redis' zunionstore.
+
+	"""
+	my_server = redis.Redis(connection_pool=POOL)
+	return my_server.zunionstore("copy_of_inactive_users",["inactive_users"])
+
+def delete_inactives_copy(delete_orig=False):
+	"""Deletes the copy of the sorted set using redis' zunionstore.
+
+	Can optionally delete the original version as well.
+	"""
+	my_server = redis.Redis(connection_pool=POOL)
+	if delete_orig:
+		pipeline1 = my_server.pipeline()
+		pipeline1.delete("copy_of_inactive_users")
+		pipeline1.delete("inactive_users")
+		pipeline1.execute()
+	else:
+		my_server.delete("copy_of_inactive_users")
+
+def bulk_sanitize_group_invite_and_membership(user_ids_list):
+	my_server = redis.Redis(connection_pool=POOL)
+	pipeline1 = my_server.pipeline()
+	for user_id in user_ids_list:
+		user_id = str(int(user_id))
+		pipeline1.delete("ug:"+user_id)
+		pipeline1.delete('ipg:'+user_id)
+		pipeline1.delete("pir:"+user_id)
+	pipeline1.execute()
+
+
+################################## export_website_feedback.py ##############################################
+
+def get_website_feedback():
+	my_server = redis.Redis(connection_pool=POOL)
+	feedback_set = "website_feedback"
+	feedback_users = my_server.smembers(feedback_set)
+	pipeline1 = my_server.pipeline()
+	complaints_with_details = []
+	for user_id in feedback_users:
+		pipeline1.hgetall("wf:"+str(user_id))
+	return pipeline1.execute()
